+++ raillabel/format/metadata.py
@@ -10,6 +10,26 @@ from importlib import metadata as importlib_metadata
 class Metadata:
     """Container for metadata information about the scene itself.
 
+    As the OpenLABEL metadata object accepts additional properties, so does this class. Any
+    properties present in the JSON will be added to the Metadata() object when read through
+    Metadata.fromdict(). Conversely, all attributes from the Metadata() object will be stored
+    into the JSON when using Metadata.asdict(). You can therefore just add attributes to the
+    Python object and have them stored.
+    Example:
+        m = Metadata.fromdict(
+            {
+                "schema_version": "1.0.0",
+                "some_additional_property": "Some Value"
+            }
+        )
+        m.another_additional_property = "Another Value"
+        m.asdict()
+        -> {
+            "schema_version": "1.0.0",
+            "some_additional_property": "Some Value",
+            "another_additional_property": "Another Value"
+        }
+
     Parameters
     ----------
     schema_version: str
@@ -46,7 +66,8 @@ class Metadata:
         Parameters
         ----------
         data_dict: dict
-            RailLabel format snippet containing the relevant data.
+            RailLabel format snippet containing the relevant data. Additional (non-defined)
+            arguments can be set and will be added as properties to Metadata.
         subschema_version: str, optional
             Version of the RailLabel subschema
 
@@ -54,20 +75,15 @@ class Metadata:
         -------
         metadata: Metadata
             Converted metadata.
-        warnings: list of str
-            List of non-critical errors, that have occurred during the conversion.
         """
 
-        return Metadata(
+        metadata = Metadata(
             schema_version=data_dict["schema_version"],
             subschema_version=subschema_version,
-            annotator=data_dict.get("annotator"),
-            file_version=data_dict.get("file_version"),
-            name=data_dict.get("name"),
-            tagged_file=data_dict.get("tagged_file"),
-            comment=data_dict.get("comment"),
             exporter_version=cls._collect_exporter_version(),
         )
+
+        return cls._set_additional_attributes(metadata, data_dict)
 
     def asdict(self) -> dict:
         """Export self as a dict compatible with the OpenLABEL schema.
@@ -76,37 +92,9 @@ class Metadata:
         -------
         dict_repr: dict
             Dict representation of this class instance.
-
-        Raises
-        ------
-        ValueError
-            if an attribute can not be converted to the type required by the OpenLabel schema.
         """
 
-        dict_repr = {"schema_version": str(self.schema_version)}
-
-        if self.annotator is not None:
-            dict_repr["annotator"] = str(self.annotator)
-
-        if self.comment is not None:
-            dict_repr["comment"] = str(self.comment)
-
-        if self.exporter_version is not None:
-            dict_repr["exporter_version"] = str(self.exporter_version)
-
-        if self.file_version is not None:
-            dict_repr["file_version"] = str(self.file_version)
-
-        if self.name is not None:
-            dict_repr["name"] = str(self.name)
-
-        if self.subschema_version is not None:
-            dict_repr["subschema_version"] = str(self.subschema_version)
-
-        if self.tagged_file is not None:
-            dict_repr["tagged_file"] = str(self.tagged_file)
-
-        return dict_repr
+        return self._remove_empty_fields(vars(self))
 
     @classmethod
     def _collect_exporter_version(cls) -> t.Optional[str]:
@@ -118,3 +106,25 @@ class Metadata:
 
         version_number_length = len(exporter_version) - len(exporter_version.split(".")[-1])
         return exporter_version[: version_number_length - 1]
+
+    @classmethod
+    def _set_additional_attributes(cls, metadata: "Metadata", data_dict: dict) -> "Metadata":
+
+        PRESET_KEYS = ["schema_version", "subschema_version", "exporter_version"]
+
+        for key, value in data_dict.items():
+            if key in PRESET_KEYS:
+                continue
+
+            is_key_a_valid_python_attribute = isinstance(key, str) and key.isidentifier()
+
+            if not is_key_a_valid_python_attribute:
+                raise KeyError(f"'{key}' is not a valid python attribute")
+
+            setattr(metadata, key, value)
+
+        return metadata
+
+    def _remove_empty_fields(self, dict_repr: dict) -> dict:
+        """Remove empty fields from a dictionary."""
+        return {k: v for k, v in dict_repr.items() if v is not None}