+++ CHANGELOG.md
@@ -57,3 +57,6 @@ Release
 ### 2.2.1
 - Fixed bug, that prevented loading a understand ai file via raillabel.load()
 - Made the project_id field in the understand ai files less restrictive
+
+# Unreleased
+- Support for additional, undefined attributes in raillabel.format.Metadata